Funciona! Solo seleccion, mutacion por inversion, y ya (cruce PMX no implementado aun)

+++ G08P1/fitness/Fitness.java
@@ -4,7 +4,7 @@ import individuo.Individuo;
 
 public interface Fitness {
 	//Obtiene el fitness del individuo
-	public double fitness(Individuo individuo);
+	public int fitness(Individuo individuo);
 	
 	public int getSize();
 }
